pyproject.toml
[tool.poetry]
name = "taskhub"
version = "0.1.0"
description = "FastAPI API for managing users and tasks"
authors = ["sheeeeeri <sheeeeeri@mail.ru>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = { extras = ["standard"], version = "^0.34.2" }
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
alembic = "^1.15.2"
python-jose = { extras = ["cryptography"], version = "^3.4.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pydantic = { extras = ["email"], version = "^2.11.4" }
python-dotenv = "^1.1.0"
pydantic-settings = "^2.9.1"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"
black = "^25.1.0"
pytest-asyncio = "^1.0.0"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
exclude = ["migrations", ".venv", "__pycache__"]
target-version = "py312"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
